+++ frontend/components/data.tsx
@@ -1,35 +1,3 @@
-import { useState } from "react";
-import reactLogo from "../public/react.svg";
-import viteLogo from "../public/vite.svg";
-import "../styles/App.css";
-
-function App() {
-  const [count, setCount] = useState(0);
-  return (
-    <>
-      <div>
-        <a href="https://vitejs.dev" target="_blank" rel="noreferrer">
-          <img src={viteLogo} className="logo" alt="Vite logo" />
-        </a>
-        <a href="https://react.dev" target="_blank" rel="noreferrer">
-          <img src={reactLogo} className="logo react" alt="React logo" />
-        </a>
-      </div>
-      <h1>Vite + React</h1>
-      <div className="card">
-        <button onClick={() => setCount((count) => count + 1)}>
-          count is {count}
-        </button>
-        <p>
-          Edit <code>src/App.tsx</code> and save to test HMR
-        </p>
-      </div>
-      <p className="read-the-docs">
-        Click on the Vite and React logos to learn more
-      </p>
-    </>
-  );
-}
 weatherType("Clouds");
 function weatherType(value: string) {
   let returnType: number = -1;
@@ -248,7 +216,6 @@ export function searchAuto(weather: string, addres: string, musicData: Song[]) {
 
 console.log(searchAuto("Snow", "大阪府", musicData));
 
-export default App;
 /* 
 天気
 Clear (晴れ)
